# Report total time on final progress bar and add metric on progress bar accuracy