--- a/src/include/duckdb/common/progress_bar/display/terminal_progress_bar_display.hpp
+++ b/src/include/duckdb/common/progress_bar/display/terminal_progress_bar_display.hpp
@@ -45,7 +45,7 @@ private:
 private:
 	static int32_t NormalizePercentage(double percentage);
 	void PrintProgressInternal(int32_t percentage, bool end);
-	double squared_progress_error = -1.0;
+	double progress_bar_error = -1.0;
 	double query_time = -1.0;
 };