--- a/src/include/duckdb/common/progress_bar/progress_bar_display.hpp
+++ b/src/include/duckdb/common/progress_bar/progress_bar_display.hpp
@@ -20,6 +20,8 @@ public:
 public:
 	virtual void Update(double percentage) = 0;
 	virtual void Finish() = 0;
+	virtual void AddInfo(const char *name, double value) {
+	}
 };
 
 } // namespace duckdb