--- a/src/main/client_context.cpp
+++ b/src/main/client_context.cpp
@@ -497,6 +497,9 @@ ClientContext::PendingPreparedStatementInternal(ClientContextLock &lock, shared_
 		}
 		active_query->progress_bar =
 		    make_uniq<ProgressBar>(executor, NumericCast<idx_t>(config.wait_time), display_create_func);
+		if (config.debug_progress_bar) {
+			active_query->progress_bar->IntializeStatCollection();
+		}
 		active_query->progress_bar->Start();
 		query_progress.Restart();
 	}